Adding FUSE 2.8 specific operation 'ioctl'.
Not sure if this will actually work yet, but it builds cleanly, and from
my reading of docs and examples, this *seems* like the right way to
implement the ioctl() call wrapper. (I'd already had this mostly
implemented and disabled, but wasn't sure it was ready for prime time.
We'll find out.)