Fix handling of XATTR_{CREATE,REPLACE} once and for all.
Okay, seems I missed a prior change that made the XATTR_CREATE and
XATTR_REPLACE functions not included when __FreeBSD__ is defined (which
also includes OS X). Also, the macros that define the symbols when
FreeBSD or NetBSD is used instead of including <sys/xattr.h> didn't
account correctly for OS X's __FreeBSD__ having a value of 10. This
corrects that.