Adding working but somewhat preliminary poll() support.
This is my current-state support for poll() (my three-quarter-assed
implementation of it anyway). I attempted to implement the poll handle
as an object, but ran into threading problems with that, so it evolved
in a slightly different direction.

Also had to add a small hack in S_fh_store_handle(), as when it
tried to use mg->mg_ptr to make the filehandle, the fsel.pl example
file ended up with all the file descriptors lost but the one. I'm not
currently sure why that's so; I've never seen that happen before, but
it seems to be threading related, as if I disable the 'threaded' flag
to Fuse::main(), it works. I think it may be an undiagnosed threading
interaction...