Change Fuse::main() unknown function handling path.
If a handler function with a name that isn't known is supplied to
Fuse::main(), instead of croak, use carp. There are legitimate situations
where a function (like one of the FUSE 2.8/2.9 specific ops) is passed,
but isn't available due to the underlying libfuse being too old to use
them, so they're just not available, and it's better to whine but not
die over them.