refactor creation/usage of perl interpreters
FUSE treats threads as light-weight objects and creates/deletes them
frequently. This leads to a new perl interpreter being created for all new
threads, but never being reclaimed/reused when a thread goes away. So, a long
running threaded mount continues to leak memory as new perl interpreters are
continuously created.

This commit is a first-pass at an alternate interpreter/thread usage pattern.
When a thread does not have an interpreter associated with it, it searches for
a currently inactive interpreter to utilize before it attempts to create one.